# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PCB Inspection Service",
    "description": "\nCompares a photo of a printed circuit board with a reference photo of the same board side,\nreturns the differing regions and collects operator feedback on them as a training dataset.\n\nAll coordinates are pixels of the uploaded image after EXIF orientation is applied.\nErrors are returned as RFC 9457 `application/problem+json`.\n",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/sessions/{session_id}/inspections": {
      "post": {
        "tags": [
          "inspections"
        ],
        "summary": "Submit Inspection",
        "description": "Queue a photo of one board side for comparison with the reference. Poll ``poll_url`` for the result.\n\nRetrying with the same ``Idempotency-Key`` returns the inspection created by the first request.",
        "operationId": "submit_inspection_api_v1_sessions__session_id__inspections_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 100,
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_inspection_api_v1_sessions__session_id__inspections_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionAccepted"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "NO_ACTIVE_REFERENCE, SESSION_CLOSED or IDEMPOTENCY_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "Upload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported image format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inspections/{inspection_id}": {
      "get": {
        "tags": [
          "inspections"
        ],
        "summary": "Get Inspection",
        "operationId": "get_inspection_api_v1_inspections__inspection_id__get",
        "parameters": [
          {
            "name": "inspection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Inspection Id"
            }
          },
          {
            "name": "wait",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 60,
              "minimum": 0,
              "description": "Long-poll: seconds to wait for a final status",
              "default": 0,
              "title": "Wait"
            },
            "description": "Long-poll: seconds to wait for a final status"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inspections/{inspection_id}/image": {
      "get": {
        "tags": [
          "inspections"
        ],
        "summary": "Get Inspection Image",
        "operationId": "get_inspection_image_api_v1_inspections__inspection_id__image_get",
        "parameters": [
          {
            "name": "inspection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Inspection Id"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ImageKind",
              "default": "original"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "image/jpeg": {},
              "image/png": {}
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inspections/{inspection_id}/defects/{defect_id}/crop": {
      "get": {
        "tags": [
          "inspections"
        ],
        "summary": "Get Crop",
        "description": "Reference and aligned test crops with identical geometry (``pair`` = side by side, reference left).",
        "operationId": "get_crop_api_v1_inspections__inspection_id__defects__defect_id__crop_get",
        "parameters": [
          {
            "name": "inspection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Inspection Id"
            }
          },
          {
            "name": "defect_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Defect Id"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/CropKind",
              "default": "pair"
            }
          },
          {
            "name": "pad",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 0,
              "description": "Context around the region, uploaded-reference pixels",
              "default": 60,
              "title": "Pad"
            },
            "description": "Context around the region, uploaded-reference pixels"
          },
          {
            "name": "height",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 32,
              "default": 400,
              "title": "Height"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "image/jpeg": {}
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with the current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inspections/{inspection_id}/defects": {
      "post": {
        "tags": [
          "inspections",
          "defects"
        ],
        "summary": "Add Manual Defect",
        "description": "Region the operator marked on the test photo. Stored as accepted, mapped to reference coordinates.",
        "operationId": "add_manual_defect_api_v1_inspections__inspection_id__defects_post",
        "parameters": [
          {
            "name": "inspection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Inspection Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualDefectIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefectOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with the current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/defects/{defect_id}/verdict": {
      "put": {
        "tags": [
          "inspections",
          "defects"
        ],
        "summary": "Set Defect Verdict",
        "description": "Operator decision; can be changed any time, every change is kept in the history.",
        "operationId": "set_defect_verdict_api_v1_defects__defect_id__verdict_put",
        "parameters": [
          {
            "name": "defect_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Defect Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerdictIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefectOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_submit_inspection_api_v1_sessions__session_id__inspections_post": {
        "properties": {
          "side": {
            "type": "integer",
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Side"
          },
          "image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Image",
            "description": "JPEG or PNG"
          },
          "board": {
            "type": "string",
            "title": "Board",
            "description": "JSON: {\"board_key\": \"...\", \"barcode\": \"...\", \"serial\": \"...\"}"
          },
          "reference_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Id"
          },
          "params": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params",
            "description": "JSON object with threshold overrides"
          },
          "captured_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          }
        },
        "type": "object",
        "required": [
          "side",
          "image",
          "board"
        ],
        "title": "Body_submit_inspection_api_v1_sessions__session_id__inspections_post"
      },
...
      "CropKind": {
        "type": "string",
        "enum": [
          "ref",
          "test",
          "pair"
        ],
        "title": "CropKind"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImageKind": {
        "type": "string",
        "enum": [
          "original",
          "aligned",
          "heatmap"
        ],
        "title": "ImageKind"
      },
      "InspectionAccepted": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/InspectionStatus"
          },
          "poll_url": {
            "type": "string",
            "title": "Poll Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "status",
          "poll_url"
        ],
        "title": "InspectionAccepted"
      },
      "InspectionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "board": {
            "$ref": "#/components/schemas/BoardIn"
          },
          "side": {
            "type": "integer",
            "title": "Side"
          },
          "reference_id": {
            "type": "string",
            "format": "uuid",
            "title": "Reference Id"
          },
          "status": {
            "$ref": "#/components/schemas/InspectionStatus"
          },
          "stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Stage"
              },
              {
                "type": "null"
              }
            ]
          },
          "superseded": {
            "type": "boolean",
            "title": "Superseded"
          },
          "image": {
            "$ref": "#/components/schemas/ImageInfo"
          },
          "quality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QualityOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "rejection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RejectionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "defects": {
            "items": {
              "$ref": "#/components/schemas/DefectOut"
            },
            "type": "array",
            "title": "Defects"
          },
          "algorithm": {
            "$ref": "#/components/schemas/AlgorithmOut"
          },
          "timings_ms": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Timings Ms"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "captured_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "session_id",
          "board",
          "side",
          "reference_id",
          "status",
          "stage",
          "superseded",
          "image",
          "quality",
          "rejection",
          "error",
          "defects",
          "algorithm",
          "timings_ms",
          "attempts",
          "captured_at",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "InspectionOut"
      },
      "InspectionStatus": {
        "type": "string",
        "enum": [
          "queued",
          "processing",
          "completed",
          "rejected",
          "failed"
        ],
        "title": "InspectionStatus"
      },
      "ManualDefectIn": {
        "properties": {
          "bbox_test": {
            "$ref": "#/components/schemas/BBoxModel"
          },
          "defect_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DefectType"
              },
              {
                "type": "null"
              }
            ]
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "operator": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "bbox_test"
        ],
        "title": "ManualDefectIn"
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail"
        ],
        "title": "Problem",
        "description": "RFC 9457 problem details."
      },
...
      "RejectionOut": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/RejectionCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "details"
        ],
        "title": "RejectionOut"
      },
...
      "Stage": {
        "type": "string",
        "enum": [
          "decode",
          "analyze",
          "save"
        ],
        "title": "Stage"
      },
...
      "Verdict": {
        "type": "string",
        "enum": [
          "pending",
          "accepted",
          "rejected"
        ],
        "title": "Verdict"
      },
      "VerdictIn": {
        "properties": {
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          },
          "defect_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DefectType"
              },
              {
                "type": "null"
              }
            ]
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "operator": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "verdict"
        ],
        "title": "VerdictIn"
      }
    }
  }
}